requirements.txt
fastapi
uvicorn[standard]
httpx
psycopg2-binary
sqlalchemy
asyncpg
redis
python-dotenv
pydantic
alt-profanity-check
groq
scikit-learn
pytest
pytest-asyncio